fix i210211 sorts folders of our style correctly
We do not zero prefix the version numbers of our files, so a naive
sort wll put 220 above 99. This extracts the version number and
sorts the folders thereby. We have adopted a couple of dfferent
styles, so this simply treats a folder with the word auto as
one of ours. Consider not using this outside of the expected context.