feat i210429 able to hide and tombstone selected, interactively
Flag -i begins an interactive session in the specified working
root. It offers the option to hide or tombstone the available
folders. It purports to allow changing the utilized screen width
and the columns used to show the folders, but doesn't actually
change the presentation meaningfully.

Adds a gradle task to run with console input. Removes main() from
Givakib.