feat i210429 honors ui commands
Able to print the available folders in the specified columns
and honor the screen character width (assumes monospace font).
This will crash if provided a combination that is too small
to render relevant output. My use case is well defined, so I'll
let that ride.
This actually calculates incorrect max values for the later
columns, but stops short of trying to render ids that aren't there.